LayoutTests/fast/mediastream/getusermedia-expected.txt

Tests webkitGetUserMedia.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS {audio:true} generated stream
PASS stream.audioTracks.length is 1
PASS stream.videoTracks.length is 0
PASS {video:true} generated stream
PASS stream.audioTracks.length is 0
PASS stream.videoTracks.length is 1
PASS {audio:true, video:true} generated stream
PASS stream.audioTracks.length is 1
PASS stream.videoTracks.length is 1
PASS successfullyParsed is true

TEST COMPLETE
